- Thêm PayPal
- Đã thêm thanh toán paypal cho học sinh
- Đã có hàm tính tiền từng tháng khóa học

+++ BaiTapTuan7/Models/tb_Course.cs
@@ -19,6 +19,6 @@ namespace BaiTapTuan7.Models
         public string Decription { get; set; }
         public string Details { get; set; }
         public Nullable<int> TeacherId { get; set; }
-        public Nullable<double> Course_Price { get; set; }
+        public Nullable<decimal> Course_Price { get; set; }
     }
 }
